sched: tune: Add prioritization logic to negative boosting
While negative values can be explicitly set from the schedtune
interface, dynamic schedtune boosts are restricted to unsigned values.
This is mainly due to conflicting logic in selecting the next boosting
values to select.

For example, say you are boosting the 'top-app' group with a value of
10. Logically, if you were to apply another value of 20, your final
value will be 20, since 20 is greater than 10 (obviously).

The trouble happens when we have an example like this: say you are
bosting the 'top-app' group with a value of 10. Now, let's apply a boost
of -10. Logically, 10 is greater than -10, so your final value will be
10. The same thing happens even when there is NO boost being applied,
since 0 is also greater than -10.

This commit adds the logic necessary to allow negative boosting values
to be chosen by the dynamic schedtune boosting algorithm. The new logic
is as follows:

If our desired boost value is greater than 0, and our desired boost
value is greater than the current boost value, apply it. If our desired
boost value is less than 0, and our desired boost value is less than the
current boost value, and the current boost is either unset or already
negative, apply it.

More logic is placed in the 'max_active_boost' method. When unboosting a
slot, the new boost value is set to the next largest boost value. This
is an issue when we have negative boosts in action, since 0 is greater
than any negative boost value. To fix that, we can detect if the
'max_boost' variable is the same as the default value. If it is, simply
return the lowest boost value instead.

This commit has been tested, and at first glance, works fine in
conjunction with CPU boosting drivers that support dynamic schedtune
boosting.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>
Signed-off-by: Dusan Uveric <dusan.uveric9@gmail.com>
Signed-off-by: nysadev <jagogardiner@gmail.com>